Check for unsupported project name
The AddIn does not support Git Xwt source
if the project is named 'Xwt'. This can not be
checked by our wizard because the name is
chosen during the last creation step out of
our control. In that case we skip the Git clone
and show an error message.

(closes #1)

(cherry picked from commit b627924680c87f1690d1bea75f2522e0af785517)